requirements.txt
python-dotenv~=1.0.0
pydub~=0.25.1
tiktoken==0.3.3
openai==0.27.4
python-telegram-bot==20.2
psycopg2==2.9.6